+++ packages/sdk/src/adaptors/goldfinch.ts
@@ -3,19 +3,25 @@ import { gqlFetcher } from '../utils';
 interface GoldfinchResult {
   data: {
     user: null | {
-      goListed: boolean;
-      type: string;
+      isGoListed: boolean;
+      type: 'BACKER' | 'CAPITAL_PROVIDER' | 'BORROWER';
     };
   };
 }
 
 export default async function getGoldfinchData(address: string) {
   const { data } = (await gqlFetcher(
-    'https://api.thegraph.com/subgraphs/name/goldfinch-eng/goldfinch',
+    'https://api.thegraph.com/subgraphs/name/goldfinch-eng/goldfinch-v2',
     `{
       user(id:"${address.toLowerCase()}"}) {
-        goListed
+        id
         type
+        isNonUsIndividual
+        isUsAccreditedIndividual
+        isUsNonAccreditedIndividual
+        isUsEntity
+        isNonUsEntity
+        isGoListed
       }
     }`
   )) as GoldfinchResult;