+++ packages/sdk/src/adaptors/labels.ts
@@ -1,15 +1,15 @@
 import { ComputeConfig } from '../types';
 import { checkComputeConfig, fetcher } from '../utils';
 
-export default async function getSybilData(
+export default async function getLabelData(
   address: string,
   computeConfig: ComputeConfig
 ) {
-  checkComputeConfig('getSybilData', computeConfig, ['CNVSEC_ID']);
+  checkComputeConfig('getLabelData', computeConfig, ['CNVSEC_ID']);
 
   const json = await fetcher(
     'GET',
-    `https://cnvsec.vercel.app/api/omnid/uniswap?id=${computeConfig.CNVSEC_ID}&address=${address}`
+    `https://cnvsec.vercel.app/api/omnid/labels?id=${computeConfig.CNVSEC_ID}&address=${address}`
   );
   return json;
 }