+++ packages/sdk/src/adaptors/quadrata.ts
@@ -8,7 +8,7 @@ interface QuadResp {
 
 export default async function getQuadrataData(address: string) {
   const { data } = (await gqlFetcher(
-    'https://api.studio.thegraph.com/query/1649/quadrata/v1.4',
+    'https://api.studio.thegraph.com/query/1649/quadrata/v1.5',
     `{
         quadrataPassports(where: {owner: "${address.toLowerCase()}"}) {
           tokenId