+++ scripts/valistRelease.js
@@ -74,8 +74,8 @@ async function publishPackage(packageName, version) {
 // 3. Run Script.
 
 const accountName = "theconvospace"
-let projectName = "react"
-let version = "0.2.2"
+let projectName = "sdk"
+let version = "0.5.10"
 
 publishPackage(`@${accountName}/${projectName}`, version).then(() => {
     process.exit(0);